% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt}
\alias{dt}
\title{Dataset for variance component analysis}
\format{
A data frame with 54 rows and 3 variables:
\describe{
  \item{Machine}{3 Levels of different machines used for production}
  \item{Metrology}{3 Levels of different metrology used for measurement}
  \item{Response}{Measured value of the characteristic under investigation}
}
}
\source{
Simulated data
}
\usage{
dt
}
\description{
Looking at the effect of Machine and Metrology on the variation in the response.
}
\keyword{datasets}
