% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt2}
\alias{dt2}
\title{Dataset for single factor variance component analysis, used in vignette}
\format{
A data frame with 36 rows and 2 variables:
\describe{
  \item{Group}{3 Level Factor}
  \item{Response}{Measured value of the characteristic under investigation}
}
}
\source{
Simulated data
}
\usage{
dt2
}
\description{
Dataset for single factor variance component analysis, used in vignette
}
\keyword{datasets}
