% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCrossProduct.r
\name{fCrossProduct}
\alias{fCrossProduct}
\title{Cross product between two vectors}
\usage{
fCrossProduct(v1, v2)
}
\arguments{
\item{v1, }{a matrix [x,y,z]}

\item{v2, }{a matrix [x,y,z]}
}
\description{
Cross product between two vectors
}
