% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fDotProduct.R
\name{fDotProduct}
\alias{fDotProduct}
\title{Dot product between two vectors}
\usage{
fDotProduct(v1, v2)
}
\arguments{
\item{v1, }{a matrix [x,y,z]}

\item{v2, }{a matrix [x,y,z]}
}
\description{
Dot product between two vectors
}
