% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGet2DRoundedBoundaryAroundPath.R
\name{fGet2DRoundedBoundaryAroundPath}
\alias{fGet2DRoundedBoundaryAroundPath}
\title{Very suboptimal code to draw a polygon around a set of x,y values - like
a bubble around a path}
\usage{
fGet2DRoundedBoundaryAroundPath(vx, vy, radius = 1)
}
\description{
Very suboptimal code to draw a polygon around a set of x,y values - like
a bubble around a path
}
