% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGet3DFromMultiple2D.R
\name{fGet3DFromMultiple2D}
\alias{fGet3DFromMultiple2D}
\title{Function to convert stereo or more vision to 3D}
\usage{
fGet3DFromMultiple2D(lProjectionList, iResolution = 1000)
}
\description{
Function to convert stereo or more vision to 3D
}
