% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetInterpolatedPointsAtDivisionPlane.r
\name{fGetInterpolatedPointsAtDivisionPlane}
\alias{fGetInterpolatedPointsAtDivisionPlane}
\title{Handles the logic needed by the mViewBeginsFromCoordinates parm in
fGetTransformedCoordinates to get the points at the dividing plane}
\usage{
fGetInterpolatedPointsAtDivisionPlane(
  mCoordinates,
  nDivisionPlaneCoefficients,
  bOriginDestinationInPositiveDirection,
  iTreatAs,
  viInputPoints = NULL
)
}
\arguments{
\item{mCoordinates}{matrix with three columns [x,y,z] with >= one rows}

\item{nDivisionPlaneCoefficients}{matrix with four columns [a,b,c,d] such that
the equation of the plan is ax + by + cz = d. This is the plane across which
the data could be spread across}

\item{bOriginDestinationInPositiveDirection}{boolean. Dictates which side
of the plane should be treated as behind and which front}

\item{iTreatAs}{Explained in mViewBeginsFromCoordinates in fGetTransformedCoordinates.}

\item{viInputPoints}{if you're tracking the order of the points already then
you can pass that vector and it will be updated. Can be left null if you don't
care for tracking the order of the points.}
}
\description{
Handles the logic needed by the mViewBeginsFromCoordinates parm in
fGetTransformedCoordinates to get the points at the dividing plane
}
