% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetOrthogonalAxes.R
\name{fGetOrthogonalAxes}
\alias{fGetOrthogonalAxes}
\title{Function to give orthogonal x and y axis vectors on a plane}
\usage{
fGetOrthogonalAxes(
  mOriginCoordinates,
  mScreenCoordinates,
  mVectorPointingUpOnScreen
)
}
\description{
Function to give orthogonal x and y axis vectors on a plane
}
