% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetPlaneAt.r
\name{fGetPlaneAt}
\alias{fGetPlaneAt}
\title{Getting plane coefficients from two vectors and a point that lie on the plane}
\usage{
fGetPlaneAt(
  mOriginCoordinates,
  mNormalVector = NULL,
  mInPlaneVector1 = NULL,
  mInPlaneVector2 = NULL
)
}
\arguments{
\item{mOriginCoordinates}{a point on the plane}

\item{mNormalVector}{vector perpedicular to the plane}

\item{mInPlaneVector1}{a vector in the plane}

\item{mInPlaneVector2}{a vector in the plane != mInPlaneVector1
You need to supply mOriginCoordinates and either mNormalVector or
mInPlaneVector1, mInPlaneVector2 to identify a plane}
}
\value{
[a,b,c,d] such that ax + by + cz = d
}
\description{
Getting plane coefficients from two vectors and a point that lie on the plane
}
