% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetProjectionsOnPlane.R
\name{fGetProjectionsOnPlane}
\alias{fGetProjectionsOnPlane}
\title{Projections of 3d points on a plane}
\usage{
fGetProjectionsOnPlane(
  mCoordinates,
  mOriginCoordinates,
  nScreenPlaneCoefficients
)
}
\arguments{
\item{mCoordinates}{matrix with three columns [x,y,z] with >= one rows}

\item{mOriginCoordinates}{three column, one row matrix specifying coordinates
of where the scene is being viewed from}

\item{nScreenPlaneCoefficients}{[a,b,c,d] such that ax + by + cz = d the
plane on which mCoordinates should be projected}
}
\description{
Projections of 3d points on a plane
}
