% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetTransformedCoordinates2.R
\name{fGetTransformedCoordinatesDiscard}
\alias{fGetTransformedCoordinatesDiscard}
\title{Transforming coordinate spaces}
\usage{
fGetTransformedCoordinatesDiscard(
  mCoordinates,
  mOriginCoordinates,
  mScreenCoordinates,
  mViewBeginsFromCoordinates = NULL,
  mZAxisVector = c(0, 0, 1),
  iTreatAs = 3
)
}
\arguments{
\item{mCoordinates}{matrix with three columns - x,y,z coorinates.
if mCoordinates is an isolated point then it is treated
as an isolated point, if it is two points, it is treated as a line segment,
but 3 or more points are treated as closed polygons unless iTreatAs = F.
Recommendation is to ensure start point is repeated as end point in mCoordinates
and iTreatAs = F. That way, where the polygon / path went behind the
screen, there will be a marker specifying that the line has to be broken there
when plotting or anywhere else}

\item{mOriginCoordinates}{three column, one row matrix specifying coordinates of where
the scene is being viewed from}

\item{mScreenCoordinates}{three column, one row matrix specifying coordinates of
where the screen on which the scene is being projected on. Also dictates the
direction along which which the scene is being viewed from}

\item{mViewBeginsFromCoordinates}{if NULL, all objects in front of origin, where
front is the direciton in which the screen coordinates are, are projected. If
a set of coordinates is given then all objects in front of that coordinate, where
front is the direciton in which the screen coordinates are, are retained, and
other objects are treated as behind the view. If this coordinate is on the opposite
side of the origin as the screen then you'll delete all objects and get an empty
array in the results}

\item{iTreatAs}{1 for isolated points, 2 for path, 3 for polygon}
}
\value{
If inputs are valid, a matrix with 2 columns for the x,y coordinates
on the screen + optional columns linking continuous stretches in front of
the screening plane which can be used in geom_path(aes(group = V3)). If invalid
inputs then you'll get a NULL.
}
\description{
Transforming coordinate spaces
}
