% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetTriangleAlongLineSegment.R
\name{fGetTriangleAlongLineSegment}
\alias{fGetTriangleAlongLineSegment}
\title{Mainly to draw arrowheads}
\usage{
fGetTriangleAlongLineSegment(
  x = 75,
  xend = 88.5,
  y = 8,
  yend = 17.6,
  baselength = 2
)
}
\arguments{
\item{baselength}{the bigger it is, the wider the base of the arrow}
}
\value{
coordinates of a triangle which form an arrowhead pointing from xy to xyend
}
\description{
Mainly to draw arrowheads
}
