% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fModelBounce.R
\name{fModelBounce}
\alias{fModelBounce}
\title{Experimental - generating random trajectories with bounes}
\usage{
fModelBounce(
  bounce_heights = c(10, 5, 2),
  bGoingUpAtStart = F,
  bGoingDownAtEnd = T,
  x = 10,
  y = 20,
  z = 3,
  endx = 40,
  endy = 30,
  endz = 2,
  ax = 0,
  ay = 0,
  az = -9.8
)
}
\description{
Experimental - generating random trajectories with bounes
}
