% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPointLineSegmentProjection.R
\name{fPointLineSegmentProjection}
\alias{fPointLineSegmentProjection}
\title{http://stackoverflow.com/questions/849211/shortest-distance-between-a-point-and-a-line-segment}
\usage{
fPointLineSegmentProjection(
  vnX,
  vnY,
  vnStartX,
  vnStartY,
  vnEndX,
  vnEndY,
  bReturnNormalisedPosition = F
)
}
\arguments{
\item{bReturnNormalisedPosition}{T - a 0 to 1 value is returned to
for where the point lies between the start point and end point, F - actual
coordinates are returned}
}
\description{
http://stackoverflow.com/questions/849211/shortest-distance-between-a-point-and-a-line-segment
}
