% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRelativeXYPositionOnPlane.R
\name{fRelativeXYPositionOnPlane}
\alias{fRelativeXYPositionOnPlane}
\title{Transforms 3d Points on a plane to the relative 2d coordinates from a specified origin}
\usage{
fRelativeXYPositionOnPlane(mCoordinates, mScreenCoordinates, mYAxis, mXAxis)
}
\arguments{
\item{mCoordinates}{the raw coordinates [x,y,z] with any number of rows}

\item{mScreenCoordinates}{a point on the plane}

\item{mYAxis}{The Y axis vector [x,y,z] with one row}

\item{mXAxis}{The X axis vector [x,y,z] with one row}
}
\description{
Transforms 3d Points on a plane to the relative 2d coordinates from a specified origin
}
