% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRemovePointsBehindDividingPlane.R
\name{fRemovePointsBehindDividingPlane}
\alias{fRemovePointsBehindDividingPlane}
\title{Handles the logic needed by the mViewBeginsFromCoordinates parm in
fGetTransformedCoordinates.}
\usage{
fRemovePointsBehindDividingPlane(
  mCoordinates,
  nDivisionPlaneCoefficients,
  bOriginDestinationInPositiveDirection,
  iTreatAs
)
}
\arguments{
\item{mCoordinates}{matrix with three columns [x,y,z] with >= one rows}

\item{nDivisionPlaneCoefficients}{matrix with four columns [a,b,c,d] such that
the equation of the plan is ax + by + cz = d. This is the plane across which
the data could be spread across}

\item{bOriginDestinationInPositiveDirection}{boolean. Dictates which side
of the plane should be treated as behind and which front}

\item{iTreatAs}{Explained in mViewBeginsFromCoordinates in fGetTransformedCoordinates.}
}
\description{
Removes unnecessary points behind the plane and keeps only those behind points
which are adjacent to a point in front for interpolation by
fGetInterpolatedPointsAtDivisionPlane
}
