% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VarTable}
\alias{VarTable}
\title{Summary table from dt, used for vignette}
\format{
A data frame with 9 rows and 5 variables:
\describe{
  \item{Machine}{3 Levels of different machines used for production}
  \item{Metrology}{3 Levels of different metrology used for measurement}
  \item{rowVariance}{Sample variance of the respone}
  \item{rowN}{Sample size of the response at each factor combination}
  \item{popVar}{Sample variance rescaled to population variance by multypling by (N-1)/N}
}
}
\source{
Simulated data
}
\usage{
VarTable
}
\description{
Summary table from dt, used for vignette
}
\keyword{datasets}
