% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGet3DTrajectoryBetweenTwoPoints.R
\name{fGet3DTrajectoryBetweenTwoPoints}
\alias{fGet3DTrajectoryBetweenTwoPoints}
\title{Calculate a trajectory between two points}
\usage{
fGet3DTrajectoryBetweenTwoPoints(
  x = 0,
  y = 0,
  z = 0,
  endx = 30,
  endy = 45,
  endz = 20,
  ax = 0,
  ay = 0,
  az = -9.8,
  t = 3,
  timesteps = 100
)
}
\description{
Calculate a trajectory between two points
}
