% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetPlaneAt.r
\name{fGetPlaneAt}
\alias{fGetPlaneAt}
\title{Getting plane coefficients from two vectors and a point that lie on the plane}
\usage{
fGetPlaneAt(
  mOriginCoordinates,
  mNormalVector = NULL,
  mVector1 = NULL,
  mVector2 = NULL
)
}
\value{
[a,b,c,d] such that ax + by + cz = d
}
\description{
Getting plane coefficients from two vectors and a point that lie on the plane
}
