% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetTextCoordinates.R
\name{fGetTextCoordinates}
\alias{fGetTextCoordinates}
\title{Will definitely have lots of edge cases, very WIP}
\usage{
fGetTextCoordinates(
  cString = NULL,
  nNormaliseOutput = T,
  bDoOutlines = F,
  cFontFamily = "Arial"
)
}
\description{
Will definitely have lots of edge cases, very WIP
}
